+++ dify_studio/kaye_cash_tracker/nodes/extract_branch/fill_prompt.py
@@ -1,5 +1,7 @@
 """
 fill the prompt template with runtime info to produce the concrete prompt
+
+todo add params
 """
 
 from datetime import datetime
@@ -17,7 +19,6 @@ def main(
     user_accounts,
     common_other_parties,
 ):  # pylint: disable=missing-function-docstring
-
     prompt = prompt_template_cache.format(
         TODAY=datetime.today().strftime("%Y-%m-%d"),
         TRANSACTIONS=transactions_array,