+++ dify_studio/kaye_event_radar/nodes/answer_branch/convert_events_to_text.py
@@ -1,9 +1,11 @@
-def main(events: list[dict]):
-    """
-    simple conversion of ``events`` array to flattened text
+"""
+simple conversion of ``events`` array to flattened text
 
 
-    :return: ``opt`` is text (typed ``str``)
-    """
+:return: ``opt`` is text (typed ``str``)
+"""
+
+
+def main(events: list[dict]):  # pylint: disable=missing-function-docstring
     opt = "\n\n".join("{}".format(event) for event in events)
     return {"output": opt}